requirements.txt
google-auth==2.17.3
google-cloud-storage==2.8.0
google-api-python-client==2.85.0